pyproject.toml
[tool.poetry]
name = "uart-over-wifi"
version = "0.1.0"
description = ""
authors = ["PymZoR <axel@mousset.me>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
python-socketio = "^5.3.0"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
